requirements.txt
Flask>=2.2
python-dotenv
twilio
tensorflow 